requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0